### 2022-05-20 #4408 @jmcbroome
The files for the paper in the docs/joss_manuscript directory in the repository. I added the paper rendering to my workflows and [it works fine](https://github.com/jmcbroome/BTE/actions/runs/2341879364). Possible this bot is looking at just the root directory? How do I fix that?
