### 2022-05-20 #4408 @Kevin-Mattheus-Moerman
@jmcbroome I think the issue is the names of the files. Can you rename it to `paper.md` and `paper.bib`?
