+++ .github/workflows/paper.yml
@@ -11,7 +11,7 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          paper-path: docs/joss_manuscript/bte.md
+          paper-path: docs/joss_manuscript/paper.md
       - name: Upload
         uses: actions/upload-artifact@v3
         with:
