### 2022-05-20 #4408 @jmcbroome
I've updated the file names and [workflow](https://github.com/jmcbroome/BTE/actions/runs/2359737669).
