### 2022-05-21 f8ef300
remove references to personal conda channel
