### 2022-05-27 #4433 @editorialbot
[REVIEW]: BTE: a Python module for pandemic-scale 2 mutation-annotated phylogenetic trees
**Submitting author:** @jmcbroome (<a href="http://orcid.org/0000-0002-5002-5156">Jakob McBroome</a>)
**Repository:** https://github.com/jmcbroome/BTE
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.2
