### 2022-06-02 8318652
improve handling of segfaults
