### 2022-06-04 #4433 @tijeco
I just wanted to double check the author list. It seems that @jmcbroome has contributed to most of the code. On the github page, I noticed @amkram was also listed as a contributor. It seems this is because of a PR they had (https://github.com/jmcbroome/BTE/pull/1) that updated the `setup.py` script with a couple of lines. I doubt that alone would be enough to warrant adding them as an author, but just wanted to check. It seems they were in the acknowledgements, perhaps that is sufficient?
