### 2022-06-15 #4433 @jmcbroome
Ah yes, sorry. @amkram helped me during initial testing phases, particularly in installing it on machines other than my development environment; he pushed some small tweaks he had to make to the setup.py to allow it to build on his machine (and on other architectures more generally) to save me the short time of reimplementing them myself. As he did not contribute any part of the actual codebase of the extension and just briefly assisted with testing, he is not listed as an author. Does that clarify the situation sufficiently @jmschrei?
