### 2022-07-28 469b10c
simplify/redo annotations function to only retrieve annotations belonging to the query node, if any exist
