### 2022-08-02 f1ddb94
remove timer wrapper on commonly repeated function calls to reduce overall print output
