### 2022-08-02 d460b4c
Add node-wise categorical entropy calculation function
