### 2022-08-03 1d3ea0b
add link to logo to readme
