+++ README.md
@@ -1,4 +1,6 @@
 # BIG TREE EXPLORER (BTE)
+<img src="bte_logo.png" width="400" height="400">
+
 Cython API for the [Mutation Annotated Tree (MAT) Online Phylogenetics Toolkit](https://github.com/yatisht/usher). 
 
 # Overview
