### 2022-08-18 b7d6bb9
add boilerplate for checking that files exist before attempting to load
