### 2022-08-18 4367849
update test suite and add handling for branch length updates when applying new mutations
