### 2022-08-19 a0d2b95
update documentation and docstrings
