### 2022-08-24 #4433 @bede
I have no concerns about this paper, with @tijeco having already raised my only concern, which has been addressed. The paper is well written, software easily installed via conda, tested on push with GitHub actions, and easily built and tested locally. The software is likely to be broadly useful and represents a significant scholarly effort.

There is basic documentation and the Jupyter notebooks accessible via Binder show nice applications of the tool and how it might actually be used.

I've PRed a one line change to the binder repo fixing a FutureWarning.
https://github.com/jmcbroome/bte-binder/pull/1

Nice work @jmcbroome et al : )
