### 2022-08-29 #4433 @tijeco
Sorry for the delay on this. I've gone through everything, and it really looks incredible! The only thing I didn't notice was a contributing guidelines section. This is a really great library, so I'm sure you will pick up more contributors!
