### 2022-08-29 d7baef6
Create CONTRIBUTING.md
