+++ CONTRIBUTING.md
@@ -1,0 +1,61 @@
+# Guidelines for Contributing
+
+We welcome contributions from the community, particularly those that fix bugs, allow BTE to work on additional architectures,
+or represent new functions/features implemented entirely within the Cython interface.
+
+## Code of Conduct
+
+Please conduct yourself respectfully. Issues may not be followed up on immediately, but should be addressed in short order.
+
+## Scope and Design
+
+BTE is, at its core, a Cython wrapper API around the C++ parsimony phylogenetics library that underlies [UShER](https://github.com/yatisht/usher), the popular tool for 
+viral phylogenetics. It is designed to work directly with the mutation-annotated parsimony phylogenetic trees produced by UShER, particularly the 
+[global SARS-CoV-2 phylogenies maintained by UCSC](http://hgdownload.soe.ucsc.edu/goldenPath/wuhCor1/UShER_SARS-CoV-2/).
+BTE also includes additional functions not found in the C++ codebase, including statistics that are too niche to justify a dedicated command line option in [matUtils](https://usher-wiki.readthedocs.io/en/latest/matUtils.html).
+It is designed for researchers and developers working on novel analyses or writing scripts performing complex operations on these trees.
+
+The scope of BTE is therefore limited to analysis of mutation-annotated phylogenetic trees. Most MATs are parsimony-based, though basic 
+likelihood trees with arbitrary branch lengths that are also mutation-annotated are supported as of recent versions. Contributions
+should therefore be focused on creating more options for working with these file types, rather than adding more expansive pipelines for tree inference or end-to-end analysis. 
+BTE is also generally designed to analyze a preexisting tree, rather than inferring a tree from raw data (a task addressed by UShER).
+
+We are particularly interested in contributions which add new features within the Cython interface (e.g. computing Robinson-Foulds distance
+between two MATrees). Functions for constructing trees from raw data are generally considered out of scope, though certainly of interest as a separate tool.
+Pure Python scripts that just import BTE are also out of scope, but are welcome as contributions to the [BTE Binder](https://github.com/jmcbroome/bte-binder). 
+
+If you are a researcher who wants to compute a particular statistic or perform a specific operation for your analysis, and you want to implement it in the more efficient Cython code,
+you are a perfect potential contributor! BTE is intended to support phylogenetics functions and operations too limited in scope to justify in the primary
+command line packages, but still of use to the research community!
+
+## Feature Suggestions
+
+If you have a suggestion for a useful additional feature, please open an issue describing the idea, and it will be discussed/approved.
+If it is within scope for BTE, the code follows formatting guidelines, and tests are included, then a follow-up pull request will be accepted.
+
+### Formatting Guidelines
+
+This module uses Python type hinting in combination with Cython typing. We therefore require that all Python functions (`def` and `cpdef`) include [type hints](https://docs.python.org/3/library/typing.html)
+in all function definitions, with the exception of hidden functions (`_func`). 
+
+The API documentation is partially autogenerated with Sphinx's `autodoc`. This works by detecting and parsing docstrings, delimited with triple ticks (\`\`\`) directly 
+following each function definition. Therefore, any functions exposed to Python (`def` and `cpdef`) must include a docstring that includes the use of, types of input, and type of output
+from the function. This ensures that new functions will be appropriately described in the documentation site. 
+
+Cython functions (`cdef`) have less strict formatting requirements, and do not require type hinting (as they often work with typed C objects already), but
+docstrings remain a useful and appreciated component of these functions. 
+
+We include a simple test script (https://github.com/jmcbroome/BTE/blob/main/run_test.py) based on python's unittest module. Please add one or more tests as necessary
+and ensure that your implemented feature does not throw an error on the small fixed test dataset before opening your pull request.
+
+## Bug Reporting
+
[…]
