### 2022-08-29 #4433 @jmcbroome
I can honestly say that I hadn't considered such guidelines, as people interested/invested enough to contribute to this kind of specialized open-source software generally reach out to me directly first with their issues or ideas. Formalized guidelines may be helpful if this repository gains popularity among the wider research community, though.

I've added a [contributing guidelines page](https://github.com/jmcbroome/BTE/commit/d7baef6282718e2d1832510889fead5420806944#diff-eca12c0a30e25b4b46522ebf89465a03ba72a03f540796c979137931d8f92055) accordingly. I hope this sufficiently addresses the reviewer's minor concern.

I would like to also thank both reviewers for their positive comments and praise!
