### 2022-08-30 #4433 @jmcbroome
I do not have a zenodo archive. Is this necessary?

The latest, I suppose- v0.8.2. Assuming that that won't require additional review steps, of course (the only substantial changes are bugfixes and a few new independent functions I've implemented, such as relative entropy).
