### 2022-08-30 #4433 @jmschrei
Yes, you need to have a doi that corresponds to the code that corresponds to your submission. Using the most recent version should be fine in this case. https://joss.readthedocs.io/en/latest/submitting.html#the-review-process
