### 2022-08-30 #4433 @jmcbroome
@jmschrei  I have uploaded the most recent source code archive with a description to Zenodo.

https://doi.org/10.5281/zenodo.7036163

Please let me know if there's anything else needed from me to complete publication!
